env, extra tab, colored alerts
