+++ packages/gui/.env.example
@@ -1,0 +1,2 @@
+ACCESSTOKEN=[YOUR ACCESS TOKEN]
+SERVER=http://localhost:3000
