+++ packages/gui/package.json
@@ -25,6 +25,7 @@
     "material-icons": "^0.7.1",
     "materialize-css": "^1.0.0",
     "mithril": "^2.0.4",
+    "mithril-materialized": "^0.19.0",
     "mithril-ui-form": "^1.1.7",
     "polylabel": "^1.1.0",
     "socket.io-client": "2.3"
@@ -34,7 +35,6 @@
     "@types/mapbox-gl": "^2.1.2",
     "@types/materialize-css": "^1.0.10",
     "@types/mithril": "^2.0.7",
-    "@types/node": "^15.12.1",
     "@types/polylabel": "^1.0.5",
     "@types/socket.io-client": "^1.4.36",
     "copy-webpack-plugin": "^9.0.0",
