+++ packages/gui/src/components/sidebars/poi-formatting.ts
